Added confirmation step before deleting categories and records

+++ components/DeleteCategoryButton.js
@@ -7,27 +7,51 @@ import Swal from 'sweetalert2'
 export default function DeleteCategoryButton({categoryId}) {
     const {user} = useContext(UserContext)
     const deleteCategory = (categoryId) => {
-        fetch(`http://localhost:4000/api/users/${user.id}/${categoryId}`, {
-            method: 'DELETE',
-            headers: {
-                Authorization: `Bearer ${localStorage.getItem('token')}`
-            }
+        Swal.fire({
+            text: 'Are you sure you want to delete this category? All records with this category will also be deleted',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Delete',
+            confirmButtonColor: '#DC3545',
+            cancelButtonText: 'Cancel'
         })
-        .then((res => res.json()))
-        .then((data) => {
-            data
-            ?   Swal.fire({
-                text: 'Deleted category!',
-                icon: 'success'
+        .then((result) => {
+            result.value
+            ?   fetch(`http://localhost:4000/api/users/${user.id}/${categoryId}`, {
+                method: 'DELETE',
+                headers: {
+                    Authorization: `Bearer ${localStorage.getItem('token')}`
+                }
                 })
-                .then((result) => {
-                    result.value
-                    ? Router.reload()
-                    : null
+                .then((res => res.json()))
+                .then((data) => {
+                    data
+                    ?   Swal.fire({
+                        text: 'Deleted category!',
+                        icon: 'success',
+                        timer: 1000,
+                        timerProgressBar: true,
+                        showConfirmButton: false
+                        })
+                        .then((result) => {
+                            result.dismiss === Swal.DismissReason.timer
+                                ? Router.reload()
+                                : Router.reload()
+                        })
+                    :   Swal.fire({
+                        text: 'Delete error!',
+                        icon: 'error',
+                        timer: 1000,
+                        timerProgressBar: true,
+                        showConfirmButton: false
+                        })
+                        .then((result) => {
+                            result.dismiss === Swal.DismissReason.timer
+                            ? Router.reload()
+                            : Router.reload()
+                        })
                 })
-            :   Swal.fire({
-                text: 'Delete error!', icon: 'error'
-                })
+            : null
         })
     }
