+++ components/DeleteRecordButton.js
@@ -7,27 +7,51 @@ import Swal from 'sweetalert2'
 export default function DeleteRecordButton({recordId}) {
     const {user} = useContext(UserContext)
     const deleteRecord = (recordId) => {
-        fetch(`http://localhost:4000/api/users/${user.id}/tr/${recordId}`, {
-            method: 'DELETE',
-            headers: {
-                Authorization: `Bearer ${localStorage.getItem('token')}`
-            }
+        Swal.fire({
+            text: 'Are you sure you want to delete this record?',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Delete',
+            confirmButtonColor: '#DC3545',
+            cancelButtonText: 'Cancel'
         })
-        .then((res => res.json()))
-        .then((data) => {
-            data
-            ?   Swal.fire({
-                text: 'Deleted record!',
-                icon: 'success'
-                })
-                .then((result) => {
-                    result.value
-                    ? Router.reload()
-                    : null
-                })
-            :   Swal.fire({
-                text: 'Delete error!', icon: 'error'
-                })
+        .then((result) => {
+            result.value
+                ?   fetch(`http://localhost:4000/api/users/${user.id}/tr/${recordId}`, {
+                    method: 'DELETE',
+                    headers: {
+                        Authorization: `Bearer ${localStorage.getItem('token')}`
+                    }
+                    })
+                    .then((res => res.json()))
+                    .then((data) => {
+                        data
+                        ?   Swal.fire({
+                            text: 'Deleted record!',
+                            icon: 'success',
+                            timer: 1000,
+                            timerProgressBar: true,
+                            showConfirmButton: false
+                            })
+                            .then((result) => {
+                                result.dismiss === Swal.DismissReason.timer
+                                ? Router.reload()
+                                : Router.reload()
+                            })
+                        :   Swal.fire({
+                            text: 'Delete error!',
+                            icon: 'error',
+                            timer: 1000,
+                            timerProgressBar: true,
+                            showConfirmButton: false
+                            })
+                            .then((result) => {
+                                result.dismiss === Swal.DismissReason.timer
+                                ? Router.reload()
+                                : Router.reload()
+                            })
+                    })
+            : null
         })
     }
